api_demo/requirements.txt
openai
json
matplotlib
numpy
networkx
z3-solver
lark
google-genai